Revert "Post input"
