+++ client/src/styling/Navigation.css
@@ -4,7 +4,7 @@
     height: 5rem;
     width: 100%;
     letter-spacing: 3px;
-    /* margin-bottom: 50px; */
+    margin-bottom: 50px;
 }
  
 .nav-link {
